+++ tests/runners/mocks.sh
@@ -1,3 +1,3 @@
 #!/bin/bash
 
-tests/docker/runners/bash/mocks.sh "$0"
+tests/docker/runners/bash/mocks.sh "$1"